+++ app/_components/uiComponents/BarGraph.tsx
