+++ app/_components/uiComponents/ProgressBar.tsx
@@ -1,4 +1,4 @@
-import styles from "../_styles/styles.module.css";
+import styles from "../../_styles/styles.module.css";
 
 export default function ProgressBar() {
     return (
