+++ app/weather/page.tsx
@@ -1,12 +1,16 @@
-import WidgetWeatherSubpage from "../_components/subComponents/WidgetWeatherSubpage";
+//import WidgetWeatherSubpage from "../_components/subComponents/WidgetWeatherSubpage";
 import BasicWidget from "../_components/layoutComponents/BasicWidget";
-import WeatherData from "../_api/weatherData";
+import WeatherData from "../_components/mainComponents/weatherData";
+import WeatherTable from "../_components/subComponents/WeatherTable";
+
+//<WidgetWeatherSubpage />
 
 export default function Page() {
     return (
         <BasicWidget className={"widget center padding-small backdrop-blur-sm min-h-[152px] lg:min-h-[202px]"}>
-            <WeatherData />
-            <WidgetWeatherSubpage />
+            <WeatherData>
+                <WeatherTable />
+            </WeatherData>
         </BasicWidget>
     );
 }
